+++ app/lib/util/native/autostart_helper.dart
@@ -108,7 +108,7 @@ Future<bool> isAutoStartHidden() async {
   }
 }
 
-const _windowsRegistryKeyValue = 'LocalSend';
+const _windowsRegistryKeyValue = 'Jet Mia';
 
 RegistryKey _getWindowsRegistryKey() {
   return Registry.openPath(