+++ app/lib/util/security_helper.dart
@@ -12,7 +12,7 @@ StoredSecurityContext generateSecurityContext([AsymmetricKeyPair? keyPair]) {
   final privateKey = keyPair.privateKey as RSAPrivateKey;
   final publicKey = keyPair.publicKey as RSAPublicKey;
   final dn = {
-    'CN': 'LocalSend User',
+    'CN': 'Jet Mia User',
     'O': '',
     'OU': '',
     'L': '',