+++ app/windows/install_msix_helper.ps1
@@ -1,1 +1,1 @@
-Add-AppxPackage .\localsend_msix_helper.msix -ExternalLocation $(Get-Location)
+Add-AppxPackage .\jet_mia_msix_helper.msix -ExternalLocation $(Get-Location)