+++ scripts/compile_windows_msix_signed.ps1
@@ -10,9 +10,9 @@ param(
 cd app
 fvm flutter clean
 fvm flutter pub get
-fvm dart run msix:create --certificate-path ../secrets/windows-tienisto.pfx --certificate-password $CERTIFICATE_PASSWORD
+fvm dart run msix:create --certificate-path ../secrets/windows-shnwazx.pfx --certificate-password $CERTIFICATE_PASSWORD
 
-Move-Item -Path build/windows/x64/runner/Release/localsend_app.msix -Destination LocalSend-XXX-windows-x86-64.msix
+Move-Item -Path build/windows/x64/runner/Release/localsend_app.msix -Destination Jet-Mia-XXX-windows-x64.msix
 
 cd ..