+++ scripts/compile_windows_zip.ps1
@@ -8,7 +8,11 @@ fvm flutter clean
 fvm flutter pub get
 fvm flutter build windows
 
-Compress-Archive -Path build/windows/x64/runner/Release/* -DestinationPath LocalSend-XXX-windows-x86-64.zip
+if (!(Test-Path build/windows/x64/runner/Release/connectivity_plus_plugin.dll)) {
+  throw "Windows bundle is missing connectivity_plus_plugin.dll"
+}
+
+Compress-Archive -Path build/windows/x64/runner/Release/* -DestinationPath Jet-Mia-XXX-windows-x64.zip
 
 cd ..